// lib/api/generated/models/OnedayClassResponse.ts
export interface OnedayClassResponse {
    /**
     * 클래스 ID
     * @type {number}
     * @memberof OnedayClassResponse
     */
    id?: number;
    /**
     * 수업 제목
     * @type {string}
     * @memberof OnedayClassResponse
     */
    name?: string;
    /**
     * 이미지 URL 목록
     * @type {Array<string>}
     * @memberof OnedayClassResponse
     */
    imageUrls?: Array<string>;
    /**
     * 수업 소개
     * @type {string}
     * @memberof OnedayClassResponse
     */
    description?: string;
    /**
     * 수업 장소
     * @type {string}
     * @memberof OnedayClassResponse
     */
    location?: string;
    /**
     * 위치 상세 안내
     * @type {string}
     * @memberof OnedayClassResponse
     */
    locationDetails?: string;
    /**
     * 준비물/재료
     * @type {string}
     * @memberof OnedayClassResponse
     */
    preparation?: string;
    /**
     * 주차 안내
     * @type {string}
     * @memberof OnedayClassResponse
     */
    parkingInfo?: string;
    /**
     * 주의사항
     * @type {string}
     * @memberof OnedayClassResponse
     */
    instructions?: string;
    /**
     * 취소 규정
     * @type {string}
     * @memberof OnedayClassResponse
     */
    cancellationPolicy?: string;
    /**
     * 클래스 코드
     * @type {string}
     * @memberof OnedayClassResponse
     */
    classCode?: string;
    /**
     * 강사 ID
     * @type {number}
     * @memberof OnedayClassResponse
     */
    instructorId?: number;
    /**
     * 링크 공유 상태
     * @type {string}
     * @memberof OnedayClassResponse
     */
    linkShareStatus?: string;
}